requirements.txt
allure-pytest
pytest
requests
python-dotenv
pydantic
playwright



